Move types to where they belong
