+++ src/interfaces/JsonSchema.ts
